requirements.txt
requests
dnspython
repr